+++ test/libsolidity/smtCheckerTests/operators/compound_bitwise_or_int_1.sol
@@ -1,0 +1,12 @@
+pragma experimental SMTChecker;
+contract C {
+    int[1][20] c;
+    function f(bool b) public {
+	    require(c[10][0] == 0);
+        if (b)
+            c[10][0] |= 1;
+        assert(c[10][0] == 0 || c[10][0] == 1);
+    }
+}
+// ----
+// Warning 1218: (174-212): Error trying to invoke SMT solver.