+++ test/libsolidity/syntaxTests/string/unicode_escape_literals_invalid_codepoint.sol
@@ -4,4 +4,4 @@ contract test {
     }
 }
 // ----
-// TypeError 6359: (86-92): Return argument type literal_string (contains invalid UTF-8 sequence at position 0) is not implicitly convertible to expected type (type of first return variable) string memory.
+// TypeError 6359: (86-92): Return argument type literal_string hex"c1" is not implicitly convertible to expected type (type of first return variable) string memory. Contains invalid UTF-8 sequence at position 0.